Change of policy to stop overly worrying about stability.
Most specifically, linear error growth is now considered to be okay, because
anything more complicated is trying to solve a problem that I don't have. None
of the current downstream applications are liable to run into precision issues
given that (i) the data values only vary by a few orders of magnitude and (ii)
we use doubles everywhere and this extends the useability of the naive methods.

So, whatever, we're going back to the simple methods. FWIW R uses the same
naive approach for all its summation methods, so I guess it's not a real
problem there, either. At least not for the datasets we deal with regularly.