Overhauled organization of functions within each category.
Each statistic now gets its own namespace with a common apply(), direct() and
Running* classes (if applicable). The row/column functions are now moved inside
this namespace and are called by_row and by_column. NaN skipping and the number
of threads are now passed via a separate statistic-specific Options class,
which avoids the need to decide on NaN handling at compile time.