Improved efficency of median calculation in the even case.
This relies on the fact that, after nth_element() is called, all
elements that are not greater than the n-th element have been moved
before the n-th element. Thus, to find the (n-1)-th element, we simply
have to take the maximum of the elements before the n-th element with
max_element(), rather than calling nth_element() again. The former is
guaranteed linear and much simpler than the latter.