Remove unnecessary checks for non-default parallelization schemes.
These checks are a historical artifact of the time when these functions
lived in tatami itself, and are largely redundant with the existing
checks for correct behavior of parallelize().

We also switch to int instead of size_t for the thread ID to avoid an
unnecessary cast inside of parallelize().