Accept a reference to a tatami::Matrix instead of a pointer.
This should be slightly safer as a reference can't be NULL. We
soft-deprecate all of the overloads that accept a pointer.

While we're at it, we also soft-deprecate the overloads that use the
default Options; it's easy to just pass '{}' in the function call, so
the overload just adds potential for confusion without much value.