Move min/max specification from template to a regular function argument.
This is simpler to understand and is consistent with the treatment of
other flags, as well as cutting down on some binary bloat. Any perf loss
should be mitigated by good branch prediction.

All the Running* classes for the range calculations now compute both the
minimum and maximum if appropriate output vectors are supplied. This is
more consistent with to the ranges::apply() interface, and users no
longer have to create two instances just to compute the range.